+++ projects/batfish/src/main/java/org/batfish/representation/cumulus/CumulusNodeConfiguration.java
@@ -5,7 +5,23 @@ import java.util.Map;
 /** A shared interfaces for the two Cumulus configuration types -- concatenated, frr */
 public interface CumulusNodeConfiguration {
 
+  public static final String LOOPBACK_INTERFACE_NAME = "lo";
+
   Map<String, IpCommunityList> getIpCommunityLists();
 
   Map<String, IpPrefixList> getIpPrefixLists();
+
+  Map<String, RouteMap> getRouteMaps();
+
+  BgpProcess getBgpProcess();
+
+  Map<String, Interface> getInterfaces();
+
+  Loopback getLoopback();
+
+  Map<String, Vrf> getVrfs();
+
+  Map<String, Vxlan> getVxlans();
+
+  OspfProcess getOspfProcess();
 }
