+++ projects/batfish/src/main/java/org/batfish/representation/cumulus/RouteMapSetAsPath.java
@@ -25,7 +25,7 @@ public class RouteMapSetAsPath implements RouteMapSet {
 
   @Nonnull
   @Override
-  public Stream<Statement> toStatements(Configuration c, CumulusNcluConfiguration vc, Warnings w) {
+  public Stream<Statement> toStatements(Configuration c, CumulusNodeConfiguration vc, Warnings w) {
     List<AsExpr> asExprs =
         _asns.stream().map(ExplicitAs::new).collect(ImmutableList.toImmutableList());
     return Stream.of(new PrependAsPath(new LiteralAsList(asExprs)));
