+++ projects/batfish/src/main/java/org/batfish/representation/cumulus/RouteMapSetIpNextHopLiteral.java
@@ -24,7 +24,7 @@ public final class RouteMapSetIpNextHopLiteral implements RouteMapSet {
 
   @Nonnull
   @Override
-  public Stream<Statement> toStatements(Configuration c, CumulusNcluConfiguration vc, Warnings w) {
+  public Stream<Statement> toStatements(Configuration c, CumulusNodeConfiguration vc, Warnings w) {
     return Stream.of(new SetNextHop(new IpNextHop(ImmutableList.of(_nextHop))));
   }
 }
