+++ projects/batfish/src/main/java/org/batfish/representation/cumulus/RouteMapSetTag.java
@@ -22,7 +22,7 @@ public class RouteMapSetTag implements RouteMapSet {
 
   @Nonnull
   @Override
-  public Stream<Statement> toStatements(Configuration c, CumulusNcluConfiguration vc, Warnings w) {
+  public Stream<Statement> toStatements(Configuration c, CumulusNodeConfiguration vc, Warnings w) {
     return Stream.of(new SetTag(new LiteralLong(_tag)));
   }
 }
